app/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
Mako==1.3.8
MarkupSafe==3.0.2
mock==5.1.0
nose==1.3.7
orjson==3.10.15
packaging==24.2
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-arango==8.1.4
python-dotenv==1.0.1
python-multipart
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
